A method for generating vectors from index-mappings.
This is analogue to `Data.Vector.generate`, whose signature is

    generate :: Int -> (Int -> a) -> Vector a

Such a function is often much more practical for building a vector
(or for conversion from other libraries' types) than having to
manually fill up tuples. It can also be considered as the dual operation
to `foldVector`, which makes it easy to (generically!) deconstruct
a SIMD vector, but previously there was no way to generically _construct_
one.

The function is not particulary efficient. Not sure if that could be
much improved yet (clearly, for best performance those operations
should generally be avoided and as much as possible be done with vector
instructions).